cleanup, build fixes and warnings fixes.

+++ linux-2.6.21.x/drivers/char/ralink_gdma.c
@@ -676,8 +676,8 @@ irqreturn_t GdmaIrqHandler(
     //GDMA_PRINT("GdmaDone Interrupt=%x\n",GdmaDoneStatus);
     //GDMA_PRINT("========================================\n");
 
-    spin_lock_irqsave(&gdma_int_lock, (unsigned long)flags);
-    
+    spin_lock_irqsave(&gdma_int_lock, flags);
+
     //UnMask error
     for(Ch=0;Ch<MAX_GDMA_CHANNEL;Ch++) {
